% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkDataCompanion.R
\docType{class}
\name{NetworkDataCompanion-class}
\alias{NetworkDataCompanion-class}
\alias{NetworkDataCompanion}
\title{NetworkDataCompanion Reference Class}
\description{
\code{NetworkDataCompanion} is a reference class that provides fields for handling
data related to TCGA and GTEx projects.
}
\details{
A package for easy and reproducible wrangling of TCGA and GTEx data.
}
\section{Fields}{

\describe{
\item{\code{TCGA_purities}}{A \code{data.frame} containing TCGA sample purity information.}

\item{\code{clinical_patient_data}}{A \code{data.frame} with clinical data for each patient.}

\item{\code{project_name}}{A \code{character} vector with the name of the project.}

\item{\code{gene_mapping}}{A \code{data.frame} that maps gene identifiers between datasets.}

\item{\code{sample_type_mapping}}{A \code{data.frame} for sample type classification and mapping.}
}}


\section{Methods}{

}

\references{
https://www.biorxiv.org/content/10.1101/2024.11.05.622163v1.abstract
}
\author{
Chen Chen (\email{chenchen9945@gmail.com}),
Viola Fanfani (\email{vfanfani@hsph.harvard.edu}),
Jonas Fischer (\email{jonas.fischer@mpi-inf.mpg.de}),
Panagiotis Mandros (\email{mandros.p@gmail.com}),
Soel Micheletti (\email{smicheletti@hsph.harvard.edu}),
Kate Hoff Shutta (\email{kshutta@hsph.harvard.edu})
}
