\name{NetworkDataCompanion}
\alias{NetworkDataCompanion}

\title{A package for easy and reproducible wrangling of TCGA and GTEx data.}

\description{Placeholder}

\references{Placeholder}

\author{
Viola Fanfani (vfanfani@hsph.harvard.edu)

Jonas Fischer (jfischer@hsph.harvard.edu)

Panagiotis Mandros (pmandros@hsph.harvard.edu)

Soel Micheletti (smicheletti@hsph.harvard.edu)

Kate Hoff Shutta (kshutta@hsph.harvard.edu)
}
