\name{NetSciDataCompanion}
\alias{NetSciDataCompanion}

\title{A package for easy and reproducible wrangling of TCGA and GTEx data.}

\description{Placeholder}

\references{Placeholder}

\author{
Jonas Fischer (jfischer@hsph.harvard.edu)

Panagiotis Mandros (pmandros@hsph.harvard.edu)

Kate Hoff Shutta (kshutta@hsph.harvard.edu)
}
