\name{probeToMeanPromoterMethylation}
\alias{probeToMeanPromoterMethylation}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{probeToMeanPromoterMethylation}
\description{Calculates the average promoter methylation within a certain window around the transcription start site (TSS), as defined by the input probe_gene_map.}
\usage{probeToMeanPromoterMethylation(methylation_betas, probe_gene_map, genesOfInterest)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{methylation_betas}{}
  \item{probe_gene_map}{}
  \item{genesOfInterest}{}
  }
}

\value{
}

\references{

}

\author{
Kate Hoff Shutta (kshutta@hsph.harvard.edu)
}
